+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 0.2.1
 -----
 
-
+* Fix bug in handling non-string column names in XLSX.
 
 0.2.0
 -----